Change UDim.Offset from int16 to int32.
Roblox increased the size at some point. Probably when UIListLayout was
added, which was the first class with a serialized UDim property.

+++ values.go
@@ -353,7 +353,7 @@ func (t ValueDouble) Copy() Value {
 
 type ValueUDim struct {
 	Scale  float32
-	Offset int16
+	Offset int32
 }
 
 func newValueUDim() Value {
